stage2_rag/requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
numpy==1.26.4
tqdm==4.66.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
scikit-learn==1.5.2
rapidfuzz==3.9.7
beautifulsoup4==4.12.3
pypdf==4.3.1
markdown==3.6
requests==2.32.3
streamlit==1.38.0
google-generativeai==0.8.3
